refactor(expose): move pricing translation out of the registry client
registry.py holds the Masumi HTTP client and the yaml-to-registry price
translation, and it reached 742 of the 750 lines this project allows a
source file. The next change to the client cannot land until something
leaves.

Pricing is the part that does not belong: it makes no request, needs no
config, and answers a question about shape rather than about a node.
Move it whole, with the constants only it uses, and re-export the names
from registry so no caller has to change.
